+++ lang/gl.js
@@ -1,0 +1,11 @@
+
+$.fullCalendar.lang("gl", {
+	buttonText: {
+		month: "Mes",
+		week: "Semana",
+		day: "Día",
+		list: "Axenda"
+	},
+	allDayHtml: "Todo<br/>o día",
+	eventLimitText: "máis"
+});