+++ src/basic/BasicView.js
@@ -112,6 +112,11 @@ var BasicView = FC.BasicView = View.extend({
 	},
 
 
+	unrenderBusinessHours: function() {
+		this.dayGrid.unrenderBusinessHours();
+	},
+
+
 	// Builds the HTML skeleton for the view.
 	// The day-grid component will render inside of a container defined by this HTML.
 	renderSkeletonHtml: function() {